requirements.txt
psycopg[binary,pool]==3.2.*
python-dotenv==1.0.*
